[ADD] draw optimisation + fix quadtree

+++ src/Engine/QuadNode.cpp
@@ -2,7 +2,7 @@
 #include "Entity.h"
 
 
-QuadNode::QuadNode(float32 x1, float32 y1, float32 x2, float32 y2, int8 d) {
+QuadNode::QuadNode(float32 x1, float32 y1, float32 x2, float32 y2, int32 d) {
 	m_bounds = { x1, y1, x2, y2 };
 	depth = d;
 }
